requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.34.28
botocore==1.34.28
Django==4.2.9
django-appconf==1.0.6
django-compressor==4.4
django-storages==1.14.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
psycopg2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
rcssmin==1.1.1
rjsmin==1.2.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==1.26.18
